# Conflicts:
#	src/main/java/com/example/javaecommerce/services/impl/ProductServiceImpl.java

+++ src/main/java/com/example/javaecommerce/services/impl/RoleServiceImpl.java
@@ -15,6 +15,8 @@ import com.example.javaecommerce.services.RoleService;
 
 import lombok.RequiredArgsConstructor;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -26,31 +28,62 @@ public class RoleServiceImpl implements RoleService {
     private final RoleRepository roleRepository;
 
     private final UserRepository userRepository;
+    private static final Logger logger = LoggerFactory.getLogger(RoleServiceImpl.class);
 
     @Override
     public List<RoleResponse> getAllRoles() {
-        List<RoleEntity> roleEntities = roleRepository.findAll();
-        return Converter.toList(roleEntities, RoleResponse.class);
+        try {
+            List<RoleEntity> roleEntities = roleRepository.findAll();
+            logger.info("Get all roles successfully !");
+            return Converter.toList(roleEntities, RoleResponse.class);
+        } catch (Exception e) {
+            logger.info("Failed to get all roles", e);
+            throw new RuntimeException("Failed to get all roles");
+        }
     }
 
     @Override
     public RoleResponse addRole(final RoleRequest role) {
-        RoleEntity roleEntity = Converter.toModel(role, RoleEntity.class);
-        roleRepository.save(roleEntity);
-        return Converter.toModel(roleEntity, RoleResponse.class);
+        try {
+            RoleEntity roleEntity = Converter.toModel(role, RoleEntity.class);
+            roleRepository.save(roleEntity);
+            logger.info("Create role successfully ! ");
+            return Converter.toModel(roleEntity, RoleResponse.class);
+        } catch (Exception e) {
+            logger.info("Failed to create role", e);
+            throw new RuntimeException("Failed to create role");
+        }
     }
 
     @Override
     public RoleResponse getRoleById(final Long roleId) {
-        RoleEntity roleEntity = roleRepository.findById(roleId)
-                .orElseThrow(() -> new EcommerceRunTimeException(ErrorCode.ID_NOT_FOUND)
-                );
-        return Converter.toModel(roleEntity, RoleResponse.class);
+        try {
+            RoleEntity roleEntity = roleRepository.findById(roleId)
+                    .orElseThrow(() -> new EcommerceRunTimeException(ErrorCode.ID_NOT_FOUND)
+                    );
+            logger.info("Get role by id successfully {} ", roleId);
+            return Converter.toModel(roleEntity, RoleResponse.class);
+        } catch (Exception e) {
+            logger.info("Failed to get role by id ", e);
+            throw new RuntimeException("Failed to get role by id");
+        }
     }
 
     @Override
     public void deleteRole(final Long roleId) {
-        roleRepository.deleteById(roleId);
+        try {
+            RoleEntity role = roleRepository
+                    .findById(roleId)
+                    .orElseThrow(() -> new EcommerceRunTimeException(ErrorCode.ID_NOT_FOUND));
+            logger.info("Cannot find role by id {} ", roleId);
+            if (role != null) {
+                roleRepository.deleteById(roleId);
+                logger.info("Delete role by id successfully {} ", roleId);
+            }
+        } catch (Exception e) {
+            logger.info("Failed to get role by id ", e);
+            throw new RuntimeException("Failed to get role by id");
+        }
     }
 
     @Override
@@ -60,13 +93,24 @@ public class RoleServiceImpl implements RoleService {
 
     @Override
     public RoleResponse addRoleForUser(final Long userId, final RoleRequest roleRequest) {
-        RoleEntity _role = userRepository.findById(userId).map(user -> {
-            String roleName = roleRequest.getName();
-            RoleEntity role = roleRepository.findByName(ERole.valueOf(roleName))
-                    .orElseThrow(() -> new EcommerceRunTimeException(ErrorCode.ID_NOT_FOUND));
-            userRepository.save(user);
-            return role;
-        }).orElseThrow(() -> new EcommerceRunTimeException(ErrorCode.ID_NOT_FOUND));
-        return Converter.toModel(_role, RoleResponse.class);
+        try {
+            RoleEntity _role = userRepository.findById(userId).map(user -> {
+                String roleName = roleRequest.getName();
+                RoleEntity role = roleRepository.findByName(ERole.valueOf(roleName))
+                        .orElseThrow(() -> new EcommerceRunTimeException(ErrorCode.ID_NOT_FOUND));
+                // get list of roles from user
+                var roles = user.getRoles();
+                // add role to role list
+                roles.add(role);
+                // save user after adding role
+                userRepository.save(user);
+                return role;
+            }).orElseThrow(() -> new EcommerceRunTimeException(ErrorCode.ID_NOT_FOUND));
+            logger.info("Add role for user successfully");
+            return Converter.toModel(_role, RoleResponse.class);
+        } catch (Exception e) {
+            logger.info("Failed to get role by id ", e);
+            throw new RuntimeException("Failed to get role by id");
+        }
     }
 }
